refactor: just send the song's ID to the background script
Since the Rust backend has been discarded for now, and since everything
is going to be self contained in the extension, the structure of how
things are done has changed a bit.

Now the content script is responsible for extracting the current playing
song's ID, and sending it to the background script, which will take care
of "liking" it in Spotify.